Conflicts:
	core/fleurdelas__define.pro

+++ core/fleurdelas__define.pro
@@ -127,19 +127,32 @@ Pro fleurDeLas::loadDataWithPath, inputFile=inputFile, _ref_extra=logMode
   ; Initializing console printing
   self.out = obj_new('consoleOutput', _extra = logMode)
 
-
+  self.sysSep = PATH_SEP()
+  
   ; Setting up a temp directory to hold temporary informations
   ; cd into the project directory
-  fleurDeLasPath = FILE_DIRNAME(ROUTINE_FILEPATH())
-  CD, fleurDeLasPath
-  CD, '../'
-  spawn, 'pwd', rootPth
-  self.rootPath = rootPth
-  ; Create a temp file
-  tempDirPath = './temp/'
-  self.tempDirPath = tempDirPath
-  command = 'mkdir '+ tempDirPath
-  spawn, command
+  fleurdelasPath = File_dirname(Routine_filepath('fleurdelas__define', /either))
+  Cd, fleurdelasPath
+  Cd, '..'
+  
+  ; Adding some support for cross plateform compatibility
+  if strlowcase(!version.os_family) eq 'unix' then begin
+    Spawn, 'pwd', rootPth
+    self.Rootpath = rootPth
+    ; Create a temp file
+    tempDirPath = rootPth + self.sysSep + 'temp'
+    self.Tempdirpath = tempDirPath
+    command = 'mkdir '+ tempDirPath
+    Spawn, command
+  endif else begin
+    Spawn, 'cd', rootPth
+    self.Rootpath = rootPth
+    ; Create a temp file
+    tempDirPath = rootPth + self.sysSep + 'temp'
+    self.Tempdirpath = tempDirPath
+    command = 'mkdir '+ tempDirPath
+    Spawn, command
+  endelse
   
   dum = self.readLAS(inputFile, header, dataStr)
   
@@ -714,10 +727,17 @@ Pro fleurDeLas::cleanup
   ; Removing the temporary files
   self.out->print,1 , 'Destroying fleurDeLas object...'
   self.out->print,1 , 'Removing temporary files...'
-  ;CD, '~/IDLWorkspace83/Saints/'
+  CD, self.rootPath
+
   ; Removing a temp file
-  command = 'rm -r '+ self.tempDirPath
+  
+  if strlowcase(!version.os_family) eq 'unix' then begin
+    command = 'rm -r '+ self.tempDirPath
   spawn, command
+  endif else begin
+    command = 'del /F /Q '+ self.tempDirPath
+  spawn, command
+  endelse
   
   
   ; Freeing all data member pointers
@@ -3456,8 +3476,9 @@ End
 ;-
 Function fleurdelas::writeLAS, id = id, output = output, selected = selected
 
-  Close, 100
-  Openw, 100, output
+
+Openw, lasLun, output, /get_lun
+
 
   self.Out->print, 1, "Writing the new file on the disk at " + Strcompress(String(output),/remove_all)
 
@@ -3516,9 +3537,11 @@ Function fleurdelas::writeLAS, id = id, output = output, selected = selected
   
   lasHeader = self->getHeaderProperty(/header)
 
-  Writeu, 100, lasHeader
+
+  Writeu, lasLun, lasHeader
   self.Out->print, 1, "Writing public header..."
-  Point_lun, -100, posHeader
+  Point_lun, -lasLun, posHeader
+
   if posHeader eq lasHeader.Headersize then self.Out->print, 1,  "Public header successfully written..."
 
 
@@ -3536,14 +3559,16 @@ Function fleurdelas::writeLAS, id = id, output = output, selected = selected
     wB = 0L
     posVlrArray = 0L
     for x=0, self.getHeaderProperty(/numberOfVLR)-1 do begin
-      Openr, 102, vlrFileID[x]
+
+      Openr, rLun, vlrFileID[x], /get_lun
       dum = Bytarr(vlrByteSize[x])
-      Readu, 102, dum
-      Point_lun, -102, r
+      Readu, rLun, dum
+      Point_lun, -rLun, r
       rB += r
-      Close, 102
-      Writeu,100,dum
-      Point_lun, -100, w
+      Close, rLun
+      Writeu, lasLun,dum
+      Point_lun, -lasLun, w
+
       if x eq 0 then begin
         wB += w-posHeader
       endif else begin
@@ -3573,15 +3598,19 @@ Function fleurdelas::writeLAS, id = id, output = output, selected = selected
     self.Out->print, 2, Strcompress(String(totalBytesWritten),/remove_all)+" bytes have been written so far, or the file header stipulates "+Strcompress(String(lasHeader.Dataoffset),/remove_all)+" bytes..."
     if lasHeader.Dataoffset-totalBytesWritten eq 2 then self.Out->print, 1, "The header is followed by 2 user-defined bytes..."
     self.Out->print, 2, "Moving "+Strcompress(String(lasHeader.Dataoffset-totalBytesWritten),/remove_all)+" bytes ahead..."
-    Point_lun, -100, actualPos
-    Point_lun, 100, actualPos + (lasHeader.Dataoffset-totalBytesWritten)
+
+    Point_lun, -lasLun, actualPos
+    Point_lun, lasLun, actualPos + (lasHeader.Dataoffset-totalBytesWritten)
+
     self.Out->print, 2, "Done... resuming writting process..."
   endelse
 
   pointSize = self->getPointSize()
   theoriticalDataBlockSize = pointSize * lasHeader.Npoints
   ;print, theoriticalDataBlockSize
-  Point_lun, -100, posBeforeDataBlock
+
+  Point_lun, -lasLun, posBeforeDataBlock
+
 
   self.Out->print, 1, "Writing points data records..."
 
@@ -3598,10 +3627,12 @@ Function fleurdelas::writeLAS, id = id, output = output, selected = selected
     endelse
   endelse
 
-  Writeu, 100, data
+
+  Writeu, lasLun, data
 
 
-  Point_lun, -100, posAfterDataBlock
+  Point_lun, -lasLun, posAfterDataBlock
+
   byte7 = posAfterDataBlock-posBeforeDataBlock
 
   self.Out->print, 1, "Checking file integrity..."
@@ -3624,22 +3655,25 @@ Function fleurdelas::writeLAS, id = id, output = output, selected = selected
     self.Out->print, 1, "Size information of the Wave Packet: ", Size(waveformBlock)
 
     self.Out->print, 1, "Writing waveforms header (EVLR header)..."
-    Writeu, 100, waveformHeader
-    Point_lun, -100, posWaveformHeader
+
+    Writeu, lasLun, waveformHeader
+    Point_lun, -lasLun, posWaveformHeader
     self.Out->print, 1, "Amount of bytes written for the waveform header: "+Strcompress(String(posWaveformHeader-posAfterDataBlock),/remove_all)
 
     self.Out->print, 1, "Writing waveforms header (EVLR header)..."
-    Writeu, 100, Byte(waveformBlock)
-    Point_lun, -100, posWaveformBlock
+    Writeu, lasLun, Byte(waveformBlock)
+    Point_lun, -lasLun, posWaveformBlock
+
     self.Out->print, 1, "Amount of bytes written for the waveform block: "+Strcompress(String(posWaveformBlock-posWaveformHeader),/remove_all)
 
   endif
 
   self.Out->print, 1, "Finilizing the file..."
-  Free_lun, 100
+
+  Free_lun, lasLun
   self.Out->print, 1, "Done."
 
-  Close, 100
+  Close,lasLun
   
   return, 1
 
@@ -3727,14 +3761,14 @@ Function fleurDeLas::dump, val, outputPath=outputPath
   
   self.out->print,1, 'Printing out ' + strcompress(string(nX)) + ' point structure into dump file...'
   
-  if keyword_set(outputPath) then openw, lun, outputPath, /APPEND, /GET_LUN else $
-                                  openw, lun, './data/ascii_dump.csv', /APPEND, /GET_LUN
+  if keyword_set(outputPath) then outputFile = outputPath else outputFile = self.rootPath + self.sysSep + 'ascii_dump.csv
+  openw, lun, outputFile, /APPEND, /GET_LUN
   
   nVal = n_elements(val)
   
   case nVal of
   
-  0: WRITE_CSV, './data/dump.csv', x, y, z
+  0: WRITE_CSV, outputFile, x, y, z
   
   1: begin
     
@@ -3939,12 +3973,14 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
 
   ; Creating a binary file that will hold the VLR Records
 
+  
+
   ; Creating a temp file that hold ALL the VLR records
   ; XXX: will need to be changed we integrated to fleurDeLas using self.tempDirPath
-  openw, 102, './temp/vlrRecords.bin'
+  vlrFilePath = self.tempDirPath + self.sysSep + 'vlrRecords.bin'
+  openw, wLun, vlrFilePath, /GET_LUN
 
-  ; Closing all lun(s) to avoid any issue
-  close, 100
+
 
 
   self.out->print,1, "Number of Variable Length Records: " + strcompress(string(fix(header.nRecords)))
@@ -3954,8 +3990,9 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
   if header.nRecords ne 0 then begin
 
     InitVRLHeader, vrlStruct
-    openr, 100, inputFile, /swap_if_big_endian
-    point_lun, 100, header.headerSize
+    openr, rLun, inputFile, /swap_if_big_endian, /get_lun
+    point_lun, rLun, header.headerSize
+
 
     ; String array containing the file name of the temp VLR files
     vlrFileArr = strarr(header.nRecords)
@@ -3969,15 +4006,17 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
 
     for w=0,header.nRecords-1,1 do begin
 
-      outputFile = strcompress('./temp/vlr_0' + string(w) + '.gkey',/REMOVE_ALL)
+      outputFile = strcompress(self.tempDirPath + self.sysSep + 'vlr_0' + string(w) + '.gkey',/REMOVE_ALL)
       vlrFileArr[w] = outputFile
 
-      readu, 100, vrlStruct
-      writeu, 102, vrlStruct
+
+      readu, rLun, vrlStruct
+      writeu, wLun, vrlStruct
 
       ; Creating a temp file that hold the nth VLR record - one file per record
-      openw, 103, outputFile
-      writeu, 103, vrlStruct
+      openw, wwLun, outputFile, /get_lun
+      writeu, wwLun, vrlStruct
+
       vlrArr[w] = ptr_new(vrlStruct)
 
       case 1 of
@@ -3993,13 +4032,16 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
             digitizerGain:0.0D,$
             digitizerOffset:0.0D $
           }
-          readu, 100, wfDescriptor
+
+          readu, rLun, wfDescriptor
+
 
           waveDescriptor = wfDescriptor
 
 
-          writeu, 102, waveDescriptor
-          writeu, 103, waveDescriptor
+          writeu, wLun, waveDescriptor
+          writeu, wwLun, waveDescriptor
+
           vlrId[w] = 1
           vlrArr[w+1] = ptr_new(waveDescriptor)
 
@@ -4018,7 +4060,9 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
             wNumberOfKeys:0US$    ;TODO to update if we add fields in there
         }
 
-        readu,100,gkdTag
+
+        readu, rLun, gkdTag
+
         ;        print, "Number of geokey:",gkdTag.wNumberOfKeys
         geoKeyHeader = gkdTag
 
@@ -4030,14 +4074,15 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
       }
 
       tempKeyEntry = replicate(sKeyEntry, gkdTag.wNumberOfKeys)
-      readu, 100, tempKeyEntry
+      readu, rLun,tempKeyEntry
 
       geoKeyArray = tempKeyEntry
 
-      writeu, 102, geoKeyHeader
-      writeu, 102, geoKeyArray
-      writeu, 103, geoKeyHeader
-      writeu, 103, geoKeyArray
+      writeu, wLun, geoKeyHeader
+      writeu, wLun, geoKeyArray
+      writeu, wwLun, geoKeyHeader
+      writeu, wwLun, geoKeyArray
+
       vlrId[w] = 2
       tempStruc = {header:gkdTag, key:tempKeyEntry}
       vlrArr[w+1] = ptr_new(tempStruc)
@@ -4048,11 +4093,13 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
     else: begin
 
       generic = bytarr(vrlStruct.recordLengthAfterHearder)
-      readu, 100,generic
+
+      readu, rLun, generic
 
 
-      writeu, 102, generic
-      writeu, 103, generic
+      writeu, wLun, generic
+      writeu, wwLun, generic
+
       vlrId[w] = 3
       vlrArr[w+1] = ptr_new(generic)
 
@@ -4060,15 +4107,18 @@ Function fleurDeLas::readVLR, inputFile, header, vlrFileArr, vlrByteSizeArr, vlr
 
   endcase
 
-  point_lun, -103, endPos
-  close, 103
+
+  point_lun, -wwLun, endPos
+  free_lun, wwLun
   vlrByteSizeArr[w] = endPos
 
 
 endfor
 
-close, 100
-close, 102
+
+free_lun, rLun
+free_lun, wLun
+
 
 endif else begin
 
@@ -4351,6 +4401,20 @@ if lasHeader.versionMinor ge 3 then dum = a->setHeaderStartWaveform(lasHeader.st
 End
 
 
+; This function creates a 3D view of the loaded data
+Function fleurdelas::view, $
+                     DUMP = DUMP    ; To be implemented
+
+  pts = self.getXYZ()
+  xyz = Transpose( [ [pts[*,0]], [pts[*,1]], [pts[*,2]] ] )
+  rgb = Transpose(self.plotGetRGBValues())
+;  rgb = Transpose([[pts.R],[pts.G],[pts.B]])/256U
+  o = Idlgrpolygon(xyz, Style=0, Vert_Colors=rgb)
+  XObjView, o, TITLE='Fleurdelas Simple Viewer"
+
+End
+
+
 
 Pro fleurDeLas__define
 
@@ -4358,8 +4422,10 @@ Pro fleurDeLas__define
   void = {fleurDeLas, $
     lasFilePath       : '',$                    ; String representing the path to the LAS file
     surveyDay         : '',$                    ; String holding the Julian Survey day - only with ARSF-NERC dataset
-    tempDirPath       : '',$                    ; String holding the temporary directory path to store temp file(s) if required
-    rootPath          : '',$                    ; String of the local path where the project is located - for relative path generation
+    tempDirPath       : '',$                    ; String holdinf the temporary directory path to store temp file(s) if required
+    rootPath          : '',$                    ; Path of the directory where the project is located
+    waveFileExt       : '',$                    ; Extention of the external waveform file, WPD | INW
+    sysSep            : '',$                    ; String of the local path where the project is located - for relative path generation
     lasHeader         : ptr_new(),$             ; Pointer to the header of the LAS file
     lasDataStr        : ptr_new(),$             ; Pointer to the Point data structure
     lasDataStrSz      : 0B,$                    ; Size in bytes of the point data structure
